Initialize project configs
Sets up the bones of the project:
- `.gitignore`
- `.gitattributes`
- `yarnrc.yml`
- `package.json`

It's worth noting, we could not use Yarn PnP, and I've opted out of Zero-Installs with Yarn 3. In other words, the project still uses `node_modules`, to work around some issues. This is reflected in our `yarnrc.yml`.

There's good information about this in yarn's docs:
- https://yarnpkg.com/getting-started/qa#which-files-should-be-gitignored
- https://yarnpkg.com/features/pnp